PLAT-10310: Fail generation when Maven/Gradle build fails
The result of the Maven/Gradle execution was not checked when generating
a Java BDK 2.0 project.

This was preventing the tests to fail if the project is not building,
which is something we want to check (if it does not compile anymore or
if the dependency version is wrong).

This will cause the generator to fail if the JDK is missing for
instance which is something we want too IMHO.