PLAT-11325: Support WDK in Symphony generator
Added WDK in the generator with corresponding specs.
For the same occasion, we corrected java/python specs to remove unuseful path.join when no dir is used.
The generated project in case of WDK is a bot ready to use with rsa credentials and a jar to run in order to execute workflows.