PLAT-11504: Generator to automatically fetch WDK latest version (#157)
In this commit, we made generator automatically fetch WDK latest version from Maven repository. We assume that both modules workflow-bot-app and workflow-language use the same version and are released together. Otherwise if this changes, then we would need to parse the api response to get the version for each module.
For the same occasion, we updated the generated build.gradle file for WDK to use DuplicatesStrategy.EXCLUDE to manage duplicated jar after download, and added maven classifier -boot- as suffix for workflow-bot-app version.