Added Yarn Audit action to the build workflow (#174)
Yarn audit action has been added to the build workflow in order to create issues when vulnerabilities are found. The scan is done on each push/pull request.
Finos scope is added in the package name. Thus, the generator dependency can be added using finos/generator-symphony (see gif below).

Pull request template has been updated in order not to mention Github issues.
For the same occasion, a typo is corrected in code of conduct markdown file.