Making the yo symphony generator a wrapper for yo @finos/symphony
Symphony generator has been moved under Finos scope, this means that now in order to call latest versions of the generator we have to type the following command "yo @finos/symphony".
Goal of this implementation is to release a new package 2.4.2 that is going to report to the user that the package as been moved and it's going to automatically use the new command.