Improve the generated project build capability (#200)
Before this commit, the generation process would stop in case the build
on the generated project fails. For instance, the package manager wrapper
cannot be downloaded due to network issue.

This commit improves the project build capability by tring to run
a the build again by using a local package manager in case there is an
issue when build fails using wrapper. In addition, no matter what the build
status is, the generation process will continue to the end.